requirements.txt
boto3==1.17.73
awsebcli==3.19.4
grpcio==1.37.1
protobuf==3.17.0
pytest==6.2.4
Flask==2.0.0
Django==3.2.3
pytube==10.8.2
moviepy==1.0.3
googleapis-common-protos==1.53.0
google-api-python-client==2.5.0
google-cloud-logging==1.15.1
google-cloud-pubsub==2.5.0
google-cloud-firestore==2.1.1
google-auth-oauthlib==0.4.4
oauth2client==4.1.3
apiclient==1.0.4
